fixed "undefined" exception thrown in lib/app.js
a check was done against res.user being null (or undefined), which
got type restricted when lib/app.js was jshinted so res.user was
required to be explicitly equal to the null object in order for
the check to pass.
the boundary of condition was changed to accomodate that.